requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-admin==2.0.2
django-allauth==0.58.2
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-excel-response2==3.0.6
django-filter==23.4
django-nine==0.2.7
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
elastic-transport==8.10.0
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
excel-base==1.0.4
idna==3.6
inflection==0.5.1
isoweek==1.3.3
msgpack==1.0.7
oauthlib==3.2.2
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.8.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
xlwt==1.3.0
